+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 ### is.codion.common.model
 - ConditionModel.dateTimePattern() now returns Optional.
 - ConditionModel.wildcard() now back to being Value based instead of Mutable.
+- TableConditionModel.persist() added, for excluding conditions when clearing the model.
 ### is.codion.common.rmi
 - ServerConfiguration.serverName() now throws exception if the server name supplier returns null or an empty string.
 - Server.serverLoad() removed, related refactoring.